Implement only user can modify his/her room

+++ base/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect
-# from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
@@ -17,6 +18,9 @@ from .forms import RoomForm
 # ]
 
 def loginPage(request):
+    
+    if request.user.is_authenicated:
+        return redirect('home')
     
     if request.method == 'POST':
         username = request.POST.get('username')
@@ -63,6 +67,8 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+
+@login_required(login_url='login')
 def create_room(request):
     form = RoomForm()   
     
@@ -75,9 +81,14 @@ def create_room(request):
     context = {'form':form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here')
+        
     
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -88,8 +99,13 @@ def update_room(request, pk):
     context = {'form' : form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def delete_room(request, pk):
     room = Room.objects.get(id= pk)
+    
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
